+++ tests/common/mock_brightness_control.h
@@ -34,7 +34,7 @@ public:
     MOCK_METHOD0(enable_autobrightness, void());
     MOCK_METHOD0(set_dim_brightness, void());
     MOCK_METHOD0(set_normal_brightness, void());
-    MOCK_METHOD1(set_normal_brightness_value, void(float));
+    MOCK_METHOD1(set_normal_brightness_value, void(double));
     MOCK_METHOD0(set_off_brightness, void());
 };
